MainPage component added. All Alg page share one component. searchURLParams added instead of useState for edit
